102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,14 +7,14 @@
  */
 int main(void)
 {
-	int count, n1 = 1, n2 = 2, next;
+	long int count, n1 = 1, n2 = 2, next;
 
-	printf("%d, %d", n1, n2);
+	printf("%ld, %ld", n1, n2);
 
 	for (count = 2; count < 50; count++)
 	{
 		next = n1 + n2;
-		printf(", %d", next);
+		printf(", %ld", next);
 		n1 = n2;
 		n2 = next;
 	}
